src/autotraits-be/requirements.txt
fastapi
uvicorn
gunicorn
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt==3.2.2
python-multipart
pandas
azure-storage-blob
black
isort
pre-commit
# --- TESTING ---
pytest
pytest-asyncio
httpx
# --- DATA ANALYSIS ---
pandas
tqdm